# gds-standalone/openapi.yaml
openapi: 3.1.0
info:
  title: Africa-first GDS API
  description: |
    Global Distribution System for African tourism properties.
    Search, book, and manage rates/availability for hotels, lodges, safari camps,
    and activities across 20+ African countries.

    ## Authentication
    Two methods supported:
    - **JWT Bearer Token**: OAuth2/OIDC from Keycloak or any compliant provider
    - **API Key**: `X-GDS-API-Key` header for server-to-server integrations

    ## Multi-Tenancy
    Set `X-GDS-Tenant-ID` header to operate within a specific tenant context.
    Omit for single-tenant deployments.

    ## Rate Limiting
    Default: 100 requests/minute per API key.
    Headers: `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`
  version: 1.0.0
  contact:
    name: TourismPay GDS Support
    email: gds@tourismpay.com
  license:
    name: Proprietary
    url: https://tourismpay.com/gds/license
servers:
  - url: http://localhost:8090
    description: Local development
  - url: https://gds-api.tourismpay.com
    description: Production
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/gds/properties:
    get:
      tags: [Properties]
      summary: List properties
      operationId: listProperties
      parameters:
        - name: country
          in: query
          schema: { type: string, example: KE }
        - name: type
          in: query
          schema: { type: string, enum: [hotel, lodge, safari_camp, resort, boutique, guesthouse, villa, apartment, activity] }
        - name: star_rating
          in: query
          schema: { type: integer, minimum: 1, maximum: 5 }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        "200":
          description: Property list
          content:
            application/json:
              schema:
                type: object
                properties:
                  properties: { type: array, items: { $ref: "#/components/schemas/Property" } }
                  total: { type: integer }
                  page: { type: integer }
                  pageSize: { type: integer }

    post:
      tags: [Properties]
      summary: Register a new property
      operationId: registerProperty
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyCreate"
      responses:
        "201":
          description: Property registered
        "400":
          description: Validation error
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-GDS-API-Key
  schemas:
    Property:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string }
        country: { type: string }
        city: { type: string }
        starRating: { type: integer }
        currency: { type: string }
        commission: { type: number }
        status: { type: string }
        amenities: { type: array, items: { type: string } }
        createdAt: { type: string, format: date-time }
    PropertyCreate:
      type: object
      required: [name, type, country, currency]
      properties:
        name: { type: string, example: Serena Safari Lodge }
        type: { type: string, enum: [hotel, lodge, safari_camp, resort, boutique, guesthouse, villa, apartment, activity] }
        country: { type: string, example: KE, description: "ISO 3166-1 alpha-2 (African countries only)" }
        region: { type: string }
        city: { type: string }
        starRating: { type: integer, minimum: 1, maximum: 5 }
        currency: { type: string, example: KES }
        commission: { type: number, description: "Agent commission percentage" }
        amenities: { type: array, items: { type: string } }
        roomTypes: { type: array, items: { $ref: "#/components/schemas/RoomType" } }
    RoomType:
      type: object
      properties:
        code: { type: string, example: DLX }
        name: { type: string, example: Deluxe Room }
        maxOccupancy: { type: integer }
        bedConfiguration: { type: string }
...